Fixed a bug with ibs distance estimate + string change for future booking

+++ Src/Server/MK.Booking.Api.Contract/Resources/DirectionInfo.cs
@@ -2,9 +2,12 @@
 {
     public class DirectionInfo : BaseDto
     {
-        public int? Distance { get; set; }
+        public double? Distance { get; set; }
+
         public double? Price { get; set; }
+
         public string FormattedDistance { get; set; }
+
 		public OrderValidationResult ValidationResult { get; set; }
 
         // Used by the webapp
